implemented selling and auto trade panel closing
selling now removes the item from the inventory and the trade screen and credits the player with the money. each transaction also checks if the sale put the player over the winning amount required and emits a signal if so.

when a player leaves the planet the trading screen is automatically closed.

updated this branch to reflect the changes on main.